Changed Varint read method to return BigInt

+++ lib/src/transaction/transaction_factory.dart
@@ -18,7 +18,7 @@ class TransactionFactory {
     var currentStep = Parsing.Version;
 
     int txInVarintByteNumber;
-    int txInVarint;
+    BigInt txInVarint;
 
     int version;
